+++ OU4/ou4_2_fib/heltal.py
@@ -10,7 +10,7 @@ class Heltal(object):
         lib.Heltal_get.restype = ctypes.c_int 
         lib.Heltal_set.argtypes = [ctypes.c_void_p,ctypes.c_int]
         lib.Heltal_fib.argtypes = [ctypes.c_void_p]
-        lib.Heltal_fib.restype = ctypes.c_int
+        lib.Heltal_fib.restype = ctypes.c_uint64
         lib.Heltal_delete.argtypes = [ctypes.c_void_p]
         self.obj = lib.Heltal_new(val)
